Removal of int32 types for gridprocessing

+++ gridprocessing/makeLayeredGrid.m
@@ -53,12 +53,12 @@ G.type = [G.type, { mfilename }];
 
 % Faces with horizontal normal
 %-----------------------------
-edges = int32(reshape(G.faces.nodes(:,1), 2, []) .');
+edges = reshape(G.faces.nodes(:,1), 2, []) .';
 dz    = G.nodes.num;
 p1    = repmat(edges(:,1), [nlayers,1])+...
-        int32(kron((0:nlayers-1)', ones(size(edges,1), 1)))*dz;
+        kron((0:nlayers-1)', ones(size(edges,1), 1))*dz;
 p2    = repmat(edges(:,2), [nlayers,1])+...
-        int32(kron((0:nlayers-1)', ones(size(edges,1), 1)))*dz;
+        kron((0:nlayers-1)', ones(size(edges,1), 1))*dz;
 hFaces = [p1, p2, p2+dz, p1+dz];
 hNumNodes = repmat(4, [size(hFaces, 1), 1]);
 
@@ -75,9 +75,9 @@ hNeighbors(bdry)=0;
 
 % Faces with vertical normal
 %----------------------------
-cn     = int32(getCellNodes(G));
+cn     = getCellNodes(G);
 vFaces = repmat(cn, [nlayers+1, 1]) + ...
-         int32(kron((0:nlayers)', ones(size(cn,1), 1)))*dz;
+         kron((0:nlayers)', ones(size(cn,1), 1))*dz;
 vNumNodes = repmat(diff(G.cells.facePos), [nlayers+1, 1]);
 
 % Neighbor relations in vertical direction.
@@ -97,7 +97,7 @@ G.faces.neighbors = [vNeighbors; hNeighbors];
 G.faces.num       = size(G.faces.neighbors,1);
 G.faces.numNodes  = [vNumNodes; hNumNodes];
 G.faces.nodes       = [vFaces(:); reshape(hFaces',[], 1)];
-pos = @(n) int32(cumsum([1; double(reshape(n, [], 1))]));
+pos = @(n) cumsum([1; double(reshape(n, [], 1))]);
 G.faces.nodePos = pos(G.faces.numNodes);
 G.cells.facePos = pos(G.cells.numFaces);
 % Cell topology from G.faces.neighbors
